pyproject.toml
[project]
name = "qwen-ocr"
version = "0.1.0"
description = "Multi-Page PDF OCR using Qwen3-VL-235B model"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai",
    "Pillow",
    "pdf2image",
    "PyPDF2",
    "tiktoken",
    "customtkinter",
    "async-tkinter-loop",
    "tkinterweb",
    "markdown-it-py[plugins]",
    "linkify-it-py>=2.0.3",
    "mdit-py-plugins>=0.5.0",
    "beautifulsoup4>=4.14.3",
    "psutil",
    "watchdog>=6.0.0",
]

[tool.pyright]
include = ["."]
exclude = ["**/__pycache__", "**/.git"]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"

[project.scripts]
qwen-ocr = "main:main"
